Rename ThePage/src/ThePage.Core/ViewModels/Bookshelf/BookShelfViewModel.cs to ThePage/src/ThePage.Core/ViewModels/BookShelf/BookShelfViewModel.cs
